update to use long for encoding

+++ Api/src/ShortUrl.Core/Extensions/Base62Extensions.cs
@@ -3,14 +3,14 @@ namespace ShortUrl.Core.Extensions;
 public static class Base62Extensions
 {
   const string base62 = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-  public static string ToBase62(this int number)
+  public static string ToBase62(this long number)
   {
     if (number == 0) return "0";
     Stack<char> stack = new Stack<char>();
 
     while (number > 0)
     {
-      stack.Push(base62[number % 62]);
+      stack.Push(base62[(int)number % 62]);
       number /= 62;
     }
