+++ Api/tests/ShortUrl.Api.Core.Tests/Base62EncodingScenarios.cs
@@ -10,7 +10,7 @@ public class Base62EncodingScenarios
     [InlineData(1, "1")]
     [InlineData(10, "A")]
     [InlineData(3844, "100")]
-    public void Should_Encode_Number_To_Base62(int number, string expected)
+    public void Should_Encode_Number_To_Base62(long number, string expected)
     {
         number.ToBase62().Should().Be(expected);
     }
